pyproject.toml
[tool.poetry]
name = "sirena-client"
version = "0.1.0"
description = "SirenaGDS Python Client"
authors = ["Utair"]
packages = [
    { include = "utair/clients/external/sirena"},
]

license = "Apache License Version 2.0, January 2004"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = {version="3.10.1"}
opentelemetry-sdk = {version=">=1.17.0"}
pydantic = {version=">2"}
xmltodict = {version="0.12.0"}
redis = {version="^4"}
asyncio = {version="*", optional = true}
aiofile = {version="*", optional = true}


[tool.poetry.extras]
async = ["asyncio", "aiofile"]



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
